Brute force the PCI space
Patch from Drew.

We were missing some of the higher buses, like fe: and ff:.  Not sure what the
relationship is between fe: and ff:.

Also, the code will still try to enumerate the entire bus behind 0:0.0, so many
of the buses are scanned twice.  I didn't want to muck with that, since it's
probably fine, and the brute force might be the wrong approach.

I tried checking 0:0.0 for multi-func, which is supposedly more PCI
controllers, but that didn't help.